VII Sem/Azure/requirements.txt
azure-functions
requests
python-dotenv>=0.19.0
azure-identity>=1.12.0
docxtpl